# Conflicts:
#	frontend/package.json
#	frontend/src/app/manage/institution-dialog/institution-dialog.component.html
#	frontend/src/app/manage/school-class-dialog/school-class-dialog.component.html
#	frontend/src/app/manage/school-year-dialog/school-year-dialog.component.html
#	frontend/src/app/manage/semester-dialog/semester-dialog.component.html
#	frontend/src/app/manage/subject-dialog/subject-dialog.component.html
#	frontend/src/app/shared/components/confirm-dialog/confirm-dialog.component.html
#	frontend/src/app/task/task-create-update-dialog/task-create-update-dialog.component.html
#	frontend/src/app/task/task-detail-component/task-detail.component.ts
#	frontend/src/app/task/task.component.html
#	frontend/yarn.lock

+++ backend/src/main/kotlin/outcobra/server/model/dto/TaskDto.kt
@@ -15,7 +15,7 @@ data class TaskDto(
         val description: String = "",
         val todoDate: LocalDate? = LocalDate.now(),
         val dueDate: LocalDate = LocalDate.now().plusDays(1),
-        val effort: Int = 0,
+        val effort: Double = 0.0,
         val progress: Int = 0) : OutcobraDto {
 
     override fun getIdentifier(): Long = id
